Extract helper function to build app in Fastfile
I decided to use an helper function instead of a private lane to
leverage Ruby's named keywords and make the interpreter fail the lane
if a caller doesn't pass all the required parameters.

Notice that the parameters defined in the lane had to become Ruby
global values (`@`) otherwise the helper wouldn't have had access to
them.

`internal_scheme` didn't need to become global, but I made it so for
consistency with the others.